Switched to using upload/download from swift shell rather than external invocation

+++ swift/bin/archive.py
@@ -1,13 +1,77 @@
 #!/usr/bin/env python3
 
 import os,sys,getopt,tarfile
-
-import swiftclient
-
-import subprocess
-
 import getpass
 
+import socket
+import optparse
+
+from swiftclient import Connection
+
+from swiftclient import shell
+from swiftclient import RequestException
+from swiftclient.exceptions import ClientException
+from swiftclient.multithreading import OutputManager
+
+# define minimum parser object to allow swiftstack shell to run 
+def shell_minimal_options():
+   parser = optparse.OptionParser()
+
+   parser.add_option('-A', '--auth', dest='auth',
+      default=os.environ.get('ST_AUTH'))
+   parser.add_option('-V', '--auth-version',
+      default=os.environ.get('ST_AUTH_VERSION',
+         (os.environ.get('OS_AUTH_VERSION','1.0'))))
+   parser.add_option('-U', '--user', dest='user',
+      default=os.environ.get('ST_USER'))
+   parser.add_option('-K', '--key', dest='key',
+      default=os.environ.get('ST_KEY'))
+ 
+   parser.add_option('--os_user_id')
+   parser.add_option('--os_user_domain_id')
+   parser.add_option('--os_user_domain_name')
+   parser.add_option('--os_tenant_id')
+   parser.add_option('--os_tenant_name')
+   parser.add_option('--os_project_id')
+   parser.add_option('--os_project_domain_id')
+   parser.add_option('--os_project_name')
+   parser.add_option('--os_project_domain_name')
+   parser.add_option('--os_service_type')
+   parser.add_option('--os_endpoint_type')
+   parser.add_option('--os_auth_token')
+   parser.add_option('--os_storage_url')
+   parser.add_option('--os_region_name')
+   
+   parser.add_option('-v', '--verbose', action='count', dest='verbose',
+       default=1, help='Print more info.')
+
+   return parser
+
+# wrapper function for swiftstack shell functions
+def sw_shell(sw_fun,*args):
+   args = ('',) + args
+   with OutputManager() as output:
+      parser = shell_minimal_options()
+      try:
+         sw_fun(parser, list(args), output)
+      except (ClientException, RequestException, socket.error) as err:
+         output.error(str(err))
+ 
+def sw_stat(*args):
+   sw_shell(shell.st_stat,*args)
+
+def sw_ls(*args):
+   sw_shell(shell.st_list,*args)
+ 
+def sw_download(*args):
+   sw_shell(shell.st_download,*args)
+ 
+def sw_upload(*args):
+   sw_shell(shell.st_upload,*args)
+ 
+def sw_post(*args):
+   sw_shell(shell.st_post,*args)
+ 
 # suffix of archive files
 tar_suffix=".tar.gz"
 
@@ -24,23 +88,15 @@ def create_tar_file(filename,src_path,file_list):
       for file in file_list:
          tar.add(os.path.join(src_path,file),file)
 
-def upload_file_to_swift(filename,swiftname,container,swift_conn):
-   # call "swift upload" if file size is greater than upload_threshold
-   upload_threshold=1000000
+def upload_file_to_swift(filename,swiftname,container):
+   sw_upload("--object-name="+swiftname,
+      "--segment-size=2147483648",
+      "--use-slo",
+      "--segment-container=.segments_testing",
+      "--header=X-Object-Meta-Uploaded-by:"+getpass.getuser(),
+      container,filename)
 
-   if os.stat(filename).st_size>upload_threshold:
-      subprocess.call(["swift","upload",
-         "--object-name="+swiftname,
-         "--segment-size=2147483648",
-         "--use-slo",
-         "--segment-container=.segments_testing",
-         "--header=X-Object-Meta-Uploaded-by:"+getpass.getuser(),
-         container,filename])
-   else:
-      with open(filename,'rb') as localfile:
-         swift_conn.put_object(container,swiftname,localfile)
-
-def archive_tar_file(src_path,file_list,container,swift_conn):
+def archive_tar_file(src_path,file_list,container):
    global tar_suffix
 
    # archive_name is name for archived object
@@ -51,18 +107,18 @@ def archive_tar_file(src_path,file_list,container,swift_conn):
    # Create local tar file 
    create_tar_file(temp_archive_name,src_path,file_list)
 
-   # Upload tar file to container as 'archive_name' using swift_conn
-   upload_file_to_swift(temp_archive_name,archive_name,container,swift_conn)
+   # Upload tar file to container as 'archive_name' 
+   upload_file_to_swift(temp_archive_name,archive_name,container)
 
    # Delete local tar file
    os.unlink(temp_archive_name)
 
 # walk directory from current
-def archive_to_swift(local_dir,container,swift_conn,no_hidden):
+def archive_to_swift(local_dir,container,no_hidden):
    for dir_name, subdir_list, file_list in os.walk(local_dir):
       rel_path=os.path.relpath(dir_name)
       if (not (no_hidden and is_hidden_dir(rel_path)) and file_list):
-         archive_tar_file(dir_name,file_list,container,swift_conn)
+         archive_tar_file(dir_name,file_list,container)
 
 # parse name into directory tree
 def create_local_path(local_dir,archive_name):
@@ -72,34 +128,47 @@ def create_local_path(local_dir,archive_name):
    if path.endswith(tar_suffix):
       path=path[:-len(tar_suffix)]
 
-   os.makedirs(path)
+   if not os.path.exists(path):
+      os.makedirs(path)
    
    return path
 
-def extract_to_local(local_dir,container,swift_conn,no_hidden):
+def create_sw_conn():
+   swift_auth=os.environ.get("ST_AUTH")
+   swift_user=os.environ.get("ST_USER")
+   swift_key=os.environ.get("ST_KEY")
+
+   if swift_auth and swift_user and swift_key:
+      return Connection(authurl=swift_auth,user=swift_user,key=swift_key)
+
+   print("Error: Swift environment not configured!")
+
+def extract_to_local(local_dir,container,no_hidden):
    global tar_suffix
 
-   try: 
-      headers,objs=swift_conn.get_container(container)
-      for obj in objs:
-         if obj['name'].endswith(tar_suffix):
-            if no_hidden and is_hidden_dir(obj['name']):
-               continue
+   swift_conn=create_sw_conn()
+   if swift_conn:
+      try: 
+         headers,objs=swift_conn.get_container(container)
+         for obj in objs:
+            if obj['name'].endswith(tar_suffix):
+               if no_hidden and is_hidden_dir(obj['name']):
+                  continue
 
-            term_path=create_local_path(local_dir,obj['name'])
+               term_path=create_local_path(local_dir,obj['name'])
 
-            # download tar file and extract into terminal directory
-            temp_file=str(os.getpid())+tar_suffix            
-            subprocess.call(["swift","download",
-               "--output="+temp_file,
-               container,obj['name']])
+               # download tar file and extract into terminal directory
+               temp_file=str(os.getpid())+tar_suffix
+               sw_download("--output="+temp_file,container,obj['name'])
 
-            with tarfile.open(temp_file,"r:gz") as tar:
-               tar.extractall(path=term_path)
+               with tarfile.open(temp_file,"r:gz") as tar:
+                  tar.extractall(path=term_path)
 
-            os.unlink(temp_file)
-   except swiftclient.exceptions.ClientException:
-      print("Error: cannot access Swift container '%s'!" % container)
+               os.unlink(temp_file)
+      except ClientException:
+         print("Error: cannot access Swift container '%s'!" % container)
+
+      swift_conn.close()
 
 def usage():
    print("archive [parameters]")
@@ -146,24 +215,11 @@ def main(argv):
    if not container:
       usage()
    else:
-      swift_auth=os.environ.get("ST_AUTH")
-      swift_user=os.environ.get("ST_USER")
-      swift_key=os.environ.get("ST_KEY")
-
-      if swift_auth and swift_user and swift_key:    
-         swift_conn=swiftclient.Connection(authurl=swift_auth,
-            user=swift_user,key=swift_key)
-
-         if swift_conn:
-            if extract:
-               extract_to_local(local_dir,container,swift_conn,no_hidden)
-            else:
-               swift_conn.put_container(container)
-               archive_to_swift(local_dir,container,swift_conn,no_hidden)
-
-            swift_conn.close()
+      if extract:
+         extract_to_local(local_dir,container,no_hidden)
       else:
-         print("Error: Missing Swift account information in environment!")
+         sw_post(container)
+         archive_to_swift(local_dir,container,no_hidden)
 
 if __name__=="__main__":
    main(sys.argv[1:])
